Cleanup unused code, rename classes, initial work on modularization
