+++ vibevoice/finetune/train_vibevoice.py
@@ -25,176 +25,8 @@ from data_vibevoice import VibeVoiceCollator, VibeVoiceDataset
 logger = logging.getLogger(__name__)
 
 
-import copy
-
-import torch
-from transformers import TrainerCallback
-
-
-class EmaCallback(TrainerCallback):
-    def __init__(self, attr_path="model.prediction_head", decay=0.999, device="cpu"):
-        """
-        attr_path: where the head lives under self.model (Trainer wraps your VibeVoiceForConditionalGeneration)
-        decay:     EMA decay (0.999 ~ stable, 0.9999 ~ very smooth, slower to adapt)
-        """
-        self.attr_path = attr_path
-        self.decay = float(decay)
-        self.device = torch.device(device)
-        self.shadow = None
-        self._orig = None  # store non-EMA weights when we swap
-
-    def _get_module(self, model):
-        # Resolve dotted path like "model.prediction_head"
-        mod = model
-        for name in self.attr_path.split("."):
-            mod = getattr(mod, name)
-        return mod
-
-    def on_train_begin(self, args, state, control, model=None, **kwargs):
-        head = self._get_module(model)
-        self.shadow = {
-            k: p.detach().to(self.device).clone() for k, p in head.state_dict().items()
-        }
-
-    def on_step_end(self, args, state, control, model=None, **kwargs):
-        if self.shadow is None:
-            return
-        head = self._get_module(model)
-        with torch.no_grad():
-            for k, v in head.state_dict().items():
-                self.shadow[k].mul_(self.decay).add_(
-                    v.detach().to(self.device), alpha=(1.0 - self.decay)
-                )
-
-    # ---- Swap helpers ----
-    def _swap_in_ema(self, model):
-        head = self._get_module(model)
-        self._orig = copy.deepcopy(head.state_dict())
-        head.load_state_dict(self.shadow, strict=False)
-
-    def _swap_back(self, model):
-        if self._orig is None:
-            return
-        head = self._get_module(model)
-        head.load_state_dict(self._orig, strict=False)
-        self._orig = None
-
-    def on_evaluate(self, args, state, control, model=None, **kwargs):
-        # use EMA during eval
-        self._swap_in_ema(model)
-
-    def on_evaluate_end(self, args, state, control, model=None, **kwargs):
-        self._swap_back(model)
-
-    def on_save(self, args, state, control, model=None, **kwargs):
-        # temporarily swap to EMA, let Trainer save, then swap back
-        self._swap_in_ema(model)
-
-    def on_save_end(self, args, state, control, model=None, **kwargs):
-        self._swap_back(model)
-
-    def on_train_end(self, args, state, control, model=None, **kwargs):
-        # final checkpoint: persist EMA
-        self._swap_in_ema(model)
-
-
-@dataclass
-class ModelArguments:
-    model_name_or_path: Optional[str] = field(
-        default="vibevoice/VibeVoice-1.5B", metadata={"help": "Path to VibeVoice base model with config.json"}
-    )
-    processor_name_or_path: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Path to processor dir (preprocessor_config.json). Defaults to model path."
-        },
-    )
-    cache_dir: Optional[str] = field(default=None)
-    freeze_acoustic_tokenizer: bool = field(default=True)
-    freeze_semantic_tokenizer: bool = field(default=True)
-    lora_r: int = field(default=8)
-    lora_alpha: int = field(default=32)
-    lora_dropout: float = field(default=0.05)
-    lora_target_modules: str = field(
-        default="q_proj,k_proj,v_proj,o_proj,gate_proj,up_proj,down_proj",
-        metadata={
-            "help": "Comma-separated list of target module names in the LLM blocks"
-        },
-    )
-    lora_wrap_diffusion_head: bool = field(
-        default=False, metadata={"help": "Wrap diffusion head with PEFT LoRA"}
-    )
-    train_diffusion_head: bool = field(
-        default=False,
-        metadata={"help": "Train diffusion prediction head (full fine-tune)"},
-    )
-    train_connectors: bool = field(
-        default=False,
-        metadata={"help": "Train acoustic/semantic connectors (full fine-tune)"},
-    )
-    layers_to_freeze: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Comma-separated indices of diffusion head layers to freeze (e.g., '0,1,5,7,8')."
-        },
-    )
-
-
-@dataclass
-class DataArguments:
-    dataset_name: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "HF dataset name or 'json' with --train_jsonl for local files"
-        },
-    )
-    dataset_config_name: Optional[str] = field(default=None)
-    train_split_name: str = field(default="train")
-    eval_split_name: Optional[str] = field(default="validation")
-    text_column_name: str = field(default="text")
-    audio_column_name: str = field(default="audio")
-    voice_prompts_column_name: Optional[str] = field(default="voice_prompts")
-    eval_split_size: float = field(default=0.0)
-    ignore_verifications: bool = field(default=False)
-    max_length: Optional[int] = field(default=None)
-    train_jsonl: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Path to local train JSONL with {text, audio, [voice_prompts]}"
-        },
-    )
-    validation_jsonl: Optional[str] = field(
-        default=None, metadata={"help": "Optional path to local validation JSONL"}
-    )
-    voice_prompt_drop_rate: float = field(
-        default=0.0,
-        metadata={
-            "help": "Probability to drop conditioning voice prompt during training (0.0 keep always, 1.0 drop always)."
-        },
-    )
-
-
-@dataclass
-class CustomTrainingArguments(HfTrainingArguments):
-    ddpm_batch_mul: int = field(default=1)
-    ce_loss_weight: float = field(default=1.0)
-    diffusion_loss_weight: float = field(default=1.0)
-    debug_ce_details: bool = field(default=False)
-    debug_ce_topk: int = field(default=5)
-    debug_ce_max_examples: int = field(default=1)
-    debug_ce_every_n_steps: int = field(default=200)
-    gradient_clipping: bool = field(
-        default=False,
-        metadata={
-            "help": "Enable gradient clipping using max_grad_norm (set via --max_grad_norm, default 1.0). When False, disables clipping by forcing max_grad_norm=0.0."
-        },
-    )
-    debug_save: bool = field(
-        default=False,
-        metadata={
-            "help": "If set, saves model components BEFORE training starts, into output_dir/debug_initial."
-        },
-    )
+from vibevoice.finetune.utils.ema_callback import EmaCallback
+from vibevoice.finetune.utils.arguments import ModelArguments, DataArguments, VibeVoiceTrainingArguments
 
 
 def build_lora_config(args: ModelArguments) -> LoraConfig:
@@ -291,7 +123,7 @@ def _patch_acoustic_encode_for_legacy_indexing(model_obj, logger_):
 
 
 def main() -> None:
-    parser = HfArgumentParser((ModelArguments, DataArguments, CustomTrainingArguments))
+    parser = HfArgumentParser((ModelArguments, DataArguments, VibeVoiceTrainingArguments))
     model_args, data_args, training_args = parser.parse_args_into_dataclasses()
 
     logging.basicConfig(
@@ -698,7 +530,7 @@ def main() -> None:
         voice_prompt_drop_rate=data_args.voice_prompt_drop_rate,
     )
 
-    class LoRADebugCallback(TrainerCallback):
+    class LoRACallback(TrainerCallback):
         def __init__(self, log_every_n_steps: int = 50):
             self.log_every_n_steps = max(1, int(log_every_n_steps))
             self.prev_param_norms: Dict[str, float] = {}
@@ -1030,83 +862,13 @@ def main() -> None:
         data_collator=data_collator,
         callbacks=[
             ema_cb,
-            LoRADebugCallback(
+            LoRACallback(
                 log_every_n_steps=(
                     int(getattr(training_args, "logging_steps", 50) or 50)
                 )
             ),
         ],
     )
-
-    # Optional debug pre-training save
-    if getattr(training_args, "debug_save", False):
-        try:
-            debug_dir = os.path.join(training_args.output_dir, "debug_initial")
-            lora_out = os.path.join(debug_dir, "lora")
-            os.makedirs(lora_out, exist_ok=True)
-            logger.info(
-                f"[debug_save] Saving initial (pre-training) model components to: {debug_dir}"
-            )
-            # language model adapters / base
-            try:
-                if hasattr(model.model.language_model, "save_pretrained"):
-                    model.model.language_model.save_pretrained(lora_out)
-            except Exception as e_lm:
-                logger.warning(f"[debug_save] Failed to save language_model: {e_lm}")
-            # diffusion head
-            try:
-                if hasattr(model.model, "prediction_head") and hasattr(
-                    model.model.prediction_head, "save_pretrained"
-                ):
-                    model.model.prediction_head.save_pretrained(
-                        os.path.join(lora_out, "diffusion_head")
-                    )
-            except Exception as e_head:
-                logger.warning(f"[debug_save] Failed to save prediction_head: {e_head}")
-            # NEW: full diffusion head state_dict as fallback
-            try:
-                ph = getattr(model.model, "prediction_head", None)
-                if ph is not None and hasattr(ph, "state_dict"):
-                    sd = ph.state_dict()
-                    torch.save(sd, os.path.join(lora_out, "diffusion_head_full.bin"))
-                    os.makedirs(os.path.join(lora_out, "diffusion_head"), exist_ok=True)
-                    torch.save(
-                        sd,
-                        os.path.join(
-                            lora_out, "diffusion_head", "diffusion_head_full.bin"
-                        ),
-                    )
-            except Exception as e:
-                logger.warning(f"[debug_save] Failed to save FULL diffusion head: {e}")
-            # connectors
-            try:
-                ac_conn = getattr(model.model, "acoustic_connector", None)
-                if ac_conn is not None:
-                    ac_dir = os.path.join(lora_out, "acoustic_connector")
-                    os.makedirs(ac_dir, exist_ok=True)
-                    torch.save(
-                        ac_conn.state_dict(), os.path.join(ac_dir, "pytorch_model.bin")
-                    )
-            except Exception as e_ac:
-                logger.warning(
-                    f"[debug_save] Failed to save acoustic_connector: {e_ac}"
-                )
-            try:
-                se_conn = getattr(model.model, "semantic_connector", None)
-                if se_conn is not None:
-                    se_dir = os.path.join(lora_out, "semantic_connector")
-                    os.makedirs(se_dir, exist_ok=True)
-                    torch.save(
-                        se_conn.state_dict(), os.path.join(se_dir, "pytorch_model.bin")
-                    )
-            except Exception as e_se:
-                logger.warning(
-                    f"[debug_save] Failed to save semantic_connector: {e_se}"
-                )
-        except Exception as e:
-            logger.warning(
-                f"[debug_save] Unexpected failure saving initial components: {e}"
-            )
 
     if getattr(training_args, "gradient_checkpointing", False):
         try:
